+++ src/index.js
@@ -1,4 +1,4 @@
-import callPop from './modules/pop.js';
+import callPop from './modules/popup.js';
 import './style.css';
 
 const apiUrl = 'https://api.tvmaze.com/shows';
@@ -13,7 +13,7 @@ const fetchMovieData = async (showId) => {
     name: data.name,
     image: data.image.medium,
     genres: data.genres,
-    itemId: showId
+    itemId: showId,
   };
 };
 
@@ -80,7 +80,12 @@ const createMovieCard = async (movieData, showId) => {
   card.appendChild(comment);
 
   comment.addEventListener('click', () => {
+    const body = document.querySelector('body');
+    // const popHTML = document.querySelector('.popHTML');
+    window.scrollTo(0, 0);
     pop.style.display = 'block';
+    body.style.overflow = 'hidden';
+    // popHTML.style.overflow = 'overflow-y: scroll';
     callPop(movieData);
   });
 
@@ -91,7 +96,7 @@ const createMovieCards = async () => {
   const response = await fetch(`${apiUrl}`);
   const showData = await response.json();
   const shows = showData.slice(0, 20);
-   
+
   shows.forEach(async (show) => {
     const movieData = await fetchMovieData(show.id);
     const movieCard = await createMovieCard(movieData, show.id);
